vlc media player batch file
When we open VLC then we have to open python code used for VLC which uses hand gestures. That was a bit irritating. 
So we came up with a batch file which will open both VLC and python code used.
One thing to keep in mind while writing batch file you have to write the address of VLC media player and python code of your laptop/pc.